process order change
