+++ src/handler/hook/post.ts
@@ -2,6 +2,6 @@ import { TextTemplate } from '../../lib/line/template'
 
 export const publishedPost = (subject: string, url: string) => {
   return TextTemplate(
-    `「${subject}」の投稿がサイトに反映されました。こちらのURLから確認できます。\n${url}/?direct=1`,
+    `「${subject}」の投稿がサイトに反映されました。こちらのURLから確認できます。\n${url}`,
   )
 }
