+++ src/handler/hook/publish_posts_handler.ts
@@ -6,12 +6,27 @@ import { loadConfig } from '../../config/config'
 import moment from 'moment'
 import { GetRecipientById } from '../../lib/firestore/recipient'
 import { GetManagerById, getManagerByLineId } from '../../lib/firestore/manager'
+import axios from 'axios'
 
 export class publishPostsHandler {
   constructor(private managerClient: Client, private recipientClient: Client) {}
 
   async handle(req: Request, res: Response) {
+    const conf = loadConfig()
+
     let posts = await getJustPublishedPosts()
+    let filteredPost = []
+    await Promise.all(
+      posts.map(async (post) => {
+        let res
+        try {
+          res = await axios.get(`${conf.frontendUrl}/post/${post.id}`)
+          if (res.status === 200) {
+            filteredPost.push(post)
+          }
+        } catch (e) {}
+      }),
+    )
     if (posts.length === 0) {
       return res.status(200).json({
         status: 'new post not found',
@@ -19,16 +34,12 @@ export class publishPostsHandler {
     }
     console.log(posts)
 
-    const conf = loadConfig()
     posts.forEach(async (post) => {
       let recipientLine = (await GetRecipientById(post.recipientId)).lineId
       let managerLine = (await GetManagerById(post.approvedBy)).lineId
-      this.recipientClient.pushMessage(recipientLine, [
-        publishedPost(post.subject, `${conf.frontendUrl}/post/${post.id}`),
-      ])
-      this.managerClient.pushMessage(managerLine, [
-        publishedPost(post.subject, `${conf.frontendUrl}/post/${post.id}`),
-      ])
+      const pageUrl = `${conf.frontendUrl}/post/${post.id}/?direct=1`
+      this.recipientClient.pushMessage(recipientLine, [publishedPost(post.subject, pageUrl)])
+      this.managerClient.pushMessage(managerLine, [publishedPost(post.subject, pageUrl)])
       post.publishedAt = moment().utcOffset(9).toDate()
       updatePost(post)
     })
